backend/requirements.txt
fastapi
uvicorn
requests
python-docx
python-dotenv
pydantic
reportlab